refactor: move filter dropdown to app level component
